requirements.txt
azure-ai-inference
azure-core
pypdf
streamlit
pandas
